Removed extraneous long ints.
git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/branches/Generalized@11082 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ src/MCMC.h
@@ -23,12 +23,12 @@ void MCMC_wrapper (int *dnumnets, int *dnedges,
                    int *maxedges);
 void MCMCSample (char *MHproposaltype, char *MHproposalpackage,
 		 double *theta, double *networkstatistics, 
-		 long int samplesize, long int burnin, 
-		 long int interval, int hammingterm, int fVerbose,
+		 int samplesize, int burnin, 
+		 int interval, int hammingterm, int fVerbose,
 	       	 Network *nwp, Model *m, DegreeBound *bd);
 void MetropolisHastings (MHproposal *MHp,
 			 double *theta, double *statistics, 
-			 long int nsteps, long int *staken,
+			 int nsteps, int *staken,
        int hammingterm,
        int fVerbose,
 			 Network *nwp, Model *m, DegreeBound *bd);
@@ -53,8 +53,8 @@ void MCMCPhase12 (int *heads, int *tails, int *dnedges,
 void MCMCSamplePhase12 (char *MHproposaltype, char *MHproposalpackage,
   double *theta, double gain, double *meanstats,
   int nphase1, int nsubphases, double *networkstatistics, 
-  long int samplesize, long int burnin, 
-  long int interval, int hammingterm, int fVerbose,
+  int samplesize, int burnin, 
+  int interval, int hammingterm, int fVerbose,
   Network *nwp, Model *m, DegreeBound *bd);
 
 #endif
